Changed float to double, short to long int, int to long int types throughout nodearray and particle array clases (SD)
git-svn-id: svn+ssh://savannah01.psi.ch/repos/gems/GEMSELEKTOR/trunk@1124 7d9feeff-4400-0410-95ed-9cb8b92a518d

+++ gems2/mods/submod/node.h
@@ -102,7 +102,7 @@ protected:
     // virtual functions for interaction with TNodeArray class (not used at TNode level)
     virtual void  setNodeArray( long int , long int*  ) { }
     virtual void  checkNodeArray( long int, long int*, const char* ) { }
-    virtual int nNodes()  const // virtual call for interaction with TNodeArray class
+    virtual long int nNodes()  const // virtual call for interaction with TNodeArray class
     { return 1; }
 
 #ifndef IPMGEMPLUGIN
